# docs/openapi.yaml
openapi: 3.0.2
info:
  title: CART MODULE
  version: 1.2.14
paths:
  /carts/attributes:
    get:
      tags:
        - cart attribute
      summary: Get List Cart Attribute
      operationId: GetListCartAttribute
      parameters:
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/AttributeFilterQuery'
        - $ref: '#/components/parameters/AttributeSortQuery'
      responses:
        '200':
          description: 'Get List Cart Attribute successfully'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AttributeModel'
        '500':
          $ref: '#/components/responses/ServerError'
      deprecated: false
  /carts/attributes/{attributeId}:
    parameters:
      - $ref: '#/components/parameters/AttributeIdParam'
    get:
      tags:
        - cart attribute
      summary: Get Cart Attribute By Id
      operationId: GetCartAttributeById
      responses:
        '200':
          description: Get Cart Attribute By Id successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/AttributeModel'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/CartAttributeNotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'
      deprecated: false
      security:
        - bearer: [ ]
...
  /carts:
    get:
      tags:
        - cart
      summary: Get Cart
      operationId: GetCart
      parameters: [ ]
      responses:
        '200':
          description: Get Cart successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/GetCartModel'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'
      deprecated: false
      security:
        - bearer: [ ]
...
  /carts/pre-checkout:
    post:
      tags:
        - cart
      summary: PreCheckout
      operationId: PreCheckout
      parameters:
        - $ref: '#/components/parameters/PreCheckoutIncludeParam'
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/PreCheckoutRequest'
                - $ref: '#/components/schemas/PreCheckoutWithDeliveryRequest'
        required: true
      responses:
        '200':
          description: PreCheckout successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        oneOf:
                          - $ref: '#/components/schemas/PreCheckoutModel'
                          - $ref: '#/components/schemas/PreCheckoutWithDeliveryModel'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'
      deprecated: false
      security:
        - bearer: [ ]
...
  /admin/carts/attributes:
    get:
      tags:
        - admin cart attribute
      summary: Get List Cart Attribute
      operationId: AdminGetListCartAttribute
      parameters:
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/AttributeFilterQuery'
        - $ref: '#/components/parameters/AttributeSortQuery'
      responses:
        '200':
          description: 'Get List Cart Attribute successfully'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AttributeModel'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'
      deprecated: false
      security:
        - bearer: [ ]
    post:
      tags:
        - admin cart attribute
      summary: Create Cart Attribute
      operationId: AdminCreateCartAttribute
      parameters: [ ]
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttributeRequest'
        required: true
      responses:
        '201':
          description: Create Attribute successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/AttributeModel'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'
      deprecated: false
      security:
        - bearer: [ ]
...
  /admin/carts/attributes/{attributeId}:
    parameters:
      - $ref: '#/components/parameters/AttributeIdParam'
    get:
      tags:
        - admin cart attribute
      summary: Get Cart Attribute By Id
      operationId: AdminGetCartAttributeById
      responses:
        '200':
          description: Get Cart Attribute By Id successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/AttributeModel'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/CartAttributeNotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'
      deprecated: false
      security:
        - bearer: [ ]
    patch:
      tags:
        - admin cart attribute
      summary: Update Cart Attribute
      operationId: AdminUpdateCartAttribute
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAttributeRequest'
        required: true
      responses:
        '204':
          description: Update Cart Attribute successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/CartAttributeNotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'
      deprecated: false
      security:
        - bearer: [ ]
    delete:
      tags:
        - admin cart attribute
      summary: Delete Cart Attribute
      operationId: AdminDeleteCartAttribute
      responses:
        '204':
          description: Delete Cart Attribute successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/CartAttributeNotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'
      deprecated: false
      security:
        - bearer: [ ]
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      description: for internal api
      type: apiKey
      name: internal-api-key
      in: header
  responses:
...
    ValidationError:
      description: Bad request
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  detail:
                    type: array
                    items:
                      type: object
                      properties:
                        property:
                          type: string
                  message:
                    type: array
                    items:
                      type: string
          example:
            error: "VALIDATION-ERR:4001"
            message: "label must be a string"
            detail:
              - property: "label must be a string"
                message:
                  - label must be a string
                  - label must be longer than or equal to 1 characters
                  - label must be shorter than or equal to 128 characters
            payloadType: body
    UnauthorizedError:
      description: Unauthorized
    ServerError:
      description: Server encountered an error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ServerErrorResponse'
    CartAttributeNotFoundError:
      description: Cart Attribute Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: CART_ATTRIBUTE:10000
            message: Cart attribute not found
...
  schemas:
    BaseResponse:
      type: object
      properties:
        data:
          type: object
...
    ServerErrorResponse:
      type: object
      properties:
        message:
          type: string
      example:
        message: Sorry our server encountered an error.
...
    AttributeType:
      type: string
      enum:
        - string
        - number
        - boolean
        - array
...
    AttributeModel:
      title: AttributeModel
      required:
        - id
        - createdAt
        - code
        - label
        - type
        - isRequired
        - isUnique
        - visibility
        - defaultValue
        - sortOrder
        - systemDefined
        - status
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          description: Tracking time create entity
          type: integer
          format: int32
        updatedAt:
          description: Tracking time update entity
          type: integer
          format: int32
        code:
          type: string
        label:
          type: string
        type:
          $ref: '#/components/schemas/AttributeType'
        isRequired:
          type: boolean
        isUnique:
          type: boolean
        visibility:
          type: boolean
        defaultValue:
          $ref: '#/components/schemas/ValueOfAttributeType'
        sortOrder:
          type: integer
          format: int32
        systemDefined:
          type: boolean
        status:
          type: boolean
      example:
        id: 63fc7ec1-43f4-41ca-a2cd-ce48fb88783f
        createdAt: 1636345541492
        updatedAt: null
        code: test
        label: Test Attribute
        type: number
        isRequired: false
        isUnique: false
        visibility: false
        defaultValue: 1
        sortOrder: 0
        systemDefined: true
        status: true
    AttributeFilter:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUIDFilterOperator'
        createdAt:
          $ref: '#/components/schemas/NumberFilterOperator'
        updatedAt:
          $ref: '#/components/schemas/NumberFilterOperator'
        code:
          $ref: '#/components/schemas/StringFilterOperator'
        label:
          $ref: '#/components/schemas/StringFilterOperator'
        type:
          $ref: '#/components/schemas/StringFilterOperator'
        isRequired:
          $ref: '#/components/schemas/BoolFilterOperator'
        isUnique:
          $ref: '#/components/schemas/BoolFilterOperator'
        visibility:
          $ref: '#/components/schemas/BoolFilterOperator'
        sortOrder:
          $ref: '#/components/schemas/NumberFilterOperator'
        systemDefined:
          $ref: '#/components/schemas/BoolFilterOperator'
        status:
          $ref: '#/components/schemas/StringFilterOperator'
    AttributeSort:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/SortOperator'
        createdAt:
          $ref: '#/components/schemas/SortOperator'
        updatedAt:
          $ref: '#/components/schemas/SortOperator'
        code:
          $ref: '#/components/schemas/SortOperator'
        label:
          $ref: '#/components/schemas/SortOperator'
        type:
          $ref: '#/components/schemas/SortOperator'
        isRequired:
          $ref: '#/components/schemas/SortOperator'
        isUnique:
          $ref: '#/components/schemas/SortOperator'
        visibility:
          $ref: '#/components/schemas/SortOperator'
        sortOrder:
          $ref: '#/components/schemas/SortOperator'
        systemDefined:
          $ref: '#/components/schemas/SortOperator'
        status:
          $ref: '#/components/schemas/SortOperator'
    CreateAttributeRequest:
      title: CreateAttributeRequest
      required:
        - code
        - label
        - type
        - isRequired
        - isUnique
        - visibility
      type: object
      properties:
        code:
          type: string
        label:
          type: string
        type:
          $ref: '#/components/schemas/AttributeType'
        isRequired:
          type: boolean
        isUnique:
          type: boolean
        visibility:
          type: boolean
        defaultValue:
          $ref: '#/components/schemas/ValueOfAttributeType'
        sortOrder:
          type: integer
          format: int32
      example:
        code: test
        label: Test Attribute
        type: number
        isRequired: false
        isUnique: false
        visibility: false
        defaultValue: 1
        sortOrder: 0
    UpdateAttributeRequest:
      title: UpdateAttributeRequest
      type: object
      properties:
        label:
          type: string
        type:
          $ref: '#/components/schemas/AttributeType'
        isRequired:
          type: boolean
        isUnique:
          type: boolean
        visibility:
          type: boolean
        defaultValue:
          $ref: '#/components/schemas/ValueOfAttributeType'
        sortOrder:
          type: integer
          format: int32
        status:
          type: string
      example:
        label: Location 1
        type: string
        isRequired: false
        isUnique: false
        visibility: false
        defaultValue: null
        sortOrder: 2
        status: disabled
...
    GetCartModel:
      title: GetCartModel
      required:
        - subtotal
        - grandTotal
        - totalSaved
        - totalItemCount
        - totalQuantity
        - cartDetail
      type: object
      properties:
        subtotal:
          type: integer
          format: int64
        grandTotal:
          type: integer
          format: int32
        totalSaved:
          type: integer
          format: int64
        totalItemCount:
          type: integer
          format: int32
        totalQuantity:
          type: integer
          format: int32
        cartDetail:
          type: array
          items:
            $ref: '#/components/schemas/CartDetailModel'
          description: ''
...
    PreCheckoutRequest:
      title: PreCheckoutRequest
      required:
        - cartDetail
        - couponIds
        - paymentMethod
      type: object
      properties:
        cartDetail:
          type: array
          items:
            $ref: '#/components/schemas/CartDetailPreCheckoutRequest'
          description: ''
        couponIds:
          type: array
          items:
            type: string
            format: uuid
          description: ''
        paymentMethod:
          type: string
          nullable: true
      example:
        cartDetail:
          - targetId: cb7137dd-156a-4551-a8f7-b388fbc6f0cb
            selectedCartItems:
              - adf15769-e0ac-49ad-b05a-e4f2ee00cf74
          - targetId: 655f5803-cf79-46b2-8d9b-d38101579855
            selectedCartItems:
              - d076b4e3-6f28-4372-9a85-58abe06ffda9
        couponIds:
          - 12557384-fe16-4ce9-87dc-86f9a6b61e41
        paymentMethod: null
    PreCheckoutWithDeliveryRequest:
      title: PreCheckoutWithDeliveryRequest
      required:
        - cartDetail
        - couponIds
        - paymentMethod
      type: object
      properties:
        cartDetail:
          type: array
          items:
            $ref: '#/components/schemas/CartDetailPreCheckoutWithDeliveryRequest'
          description: ''
        couponIds:
          type: array
          items:
            type: string
            format: uuid
          description: ''
        paymentMethod:
          type: string
          nullable: true
        recipient:
          $ref: '#/components/schemas/RecipientRequest'
      example:
        cartDetail:
          - targetId: cb7137dd-156a-4551-a8f7-b388fbc6f0cb
            selectedCartItems:
              - adf15769-e0ac-49ad-b05a-e4f2ee00cf74
          - targetId: 655f5803-cf79-46b2-8d9b-d38101579855
            selectedCartItems:
              - d076b4e3-6f28-4372-9a85-58abe06ffda9
        couponIds:
          - 12557384-fe16-4ce9-87dc-86f9a6b61e41
        paymentMethod: null
...
    PreCheckoutModel:
      title: PreCheckoutModel
      required:
        - checkoutPriceData
        - cartDetail
        - promotionData
        - totalQuantity
        - totalItemCount
        - couponUsedQuantity
      type: object
      properties:
        checkoutPriceData:
          $ref: '#/components/schemas/CheckoutPriceDataModel'
        cartDetail:
          type: array
          items:
            $ref: '#/components/schemas/CartDetailReCheckoutModel'
          description: ''
        promotionData:
          $ref: '#/components/schemas/PromotionDataModel'
        totalQuantity:
          type: integer
          format: int32
        totalItemCount:
          type: integer
          format: int32
        couponUsedQuantity:
          type: integer
          format: int32
...
    PreCheckoutWithDeliveryModel:
      title: PreCheckoutWithDeliveryModel
      required:
        - checkoutPriceData
        - cartDetail
        - promotionData
        - totalQuantity
        - totalItemCount
        - couponUsedQuantity
      type: object
      properties:
        checkoutPriceData:
          $ref: '#/components/schemas/CheckoutPriceDataWithDeliveryModel'
        cartDetail:
          type: array
          items:
            $ref: '#/components/schemas/CartDetailReCheckoutWithDeliveryModel'
          description: ''
        promotionData:
          $ref: '#/components/schemas/PromotionDataModel'
        totalQuantity:
          type: integer
          format: int32
        totalItemCount:
          type: integer
          format: int32
        couponUsedQuantity:
          type: integer
          format: int32
...
  parameters:
    OffsetParam:
      in: query
      name: offset
      schema:
        type: integer
        minimum: 0
        maximum: 100000
        default: 0
      required: false
      description: Current offset
    LimitParam:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100000
        default: 10
      required: false
      description: Max length list data
...
    AttributeIdParam:
      in: path
      name: attributeId
      schema:
        type: string
        format: uuid
        example: 91624780-592a-4c2e-af96-b2f0fa455ef4
      required: true
      description: attribute id is uuid
    AttributeFilterQuery:
      in: query
      name: filter
      schema:
        type: object
        allOf:
          - $ref: '#/components/schemas/AttributeFilter'
      style: deepObject
      explode: true
      description: (Schema AttributeFilter) Options for filtering the results
      required: false
    AttributeSortQuery:
      in: query
      name: sort
      schema:
        type: object
        allOf:
          - $ref: '#/components/schemas/AttributeSort'
      style: deepObject
      explode: true
      description: (Schema AttributeSort) Options for sorting the results
      required: false
...
    PreCheckoutIncludeParam:
      in: query
      name: includes
      schema:
        type: string
        enum:
          - delivery
        nullable: true
      required: false
      description: Max length list data